chore:將 api 內的 type 移入到 type資料夾內
